stockflow/requirements.txt
fastapi
uvicorn
jinja2
bcrypt
python-jose[cryptography]
python-multipart
pyyaml
